redo dir structure
