+++ brackets/core
@@ -1,0 +1,1 @@
+../core
