+++ browser/index.html
@@ -30,7 +30,7 @@
     <!-- CSS/LESS -->
 
     <link rel="stylesheet" href="lib/bootstrap.min.css">
-    <link rel="stylesheet/less" href="styles/ewf-browser.less">
+    <link rel="stylesheet/less" href="../core/styles/ewf-browser.less">
     <script src="lib/less-1.3.0.min.js"></script>
     
     <!-- JavaScript -->
