+++ browser/lib/i18n.js
