+++ browser/lib/mustache.js
