+++ browser/lib/require.js
