+++ browser/lib/text.js
