+++ core/nls/de/strings.js
