+++ core/nls/fr/strings.js
