+++ core/nls/root/strings.js
