+++ core/nls/strings.js
