+++ core/styles/ewf-browser.less
