+++ server/server.js
@@ -42,7 +42,7 @@
     
     var app = connect()
         .use(connect.logger('dev'))
-        .use(connect.theStatic('../'))
+        .use(connect.theStatic('../browser/'))
         .use(function (req, res) {
             if (req.url.indexOf('/proxy/') === 0) {
                 var path = api_path_prefix + req.url.substr('/proxy/'.length);
